requirements.txt
Flask==3.0.2
Flask-Mail==0.10.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
inflect==7.3.1
pdfkit==1.0.0
reportlab==4.2.2
pypdf==4.3.1
PyMySQL==1.1.0
Werkzeug==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
